Move ButlerConfig to its own file
